+++ src/app/components/users-module/users/users.component.ts
@@ -135,7 +135,16 @@ export class UsersComponent implements OnInit {
   }
 
   // Función para eliminar un usuario
-  onDeleteAction(event: any) {
+  async onDeleteAction(event: any) {
+    const confirmed = await this.auxService.AlertConfirmation(
+      'Are you sure you want to delete this record?',
+      'This action cannot be undone.',
+      'Yes, delete it!'
+    );
+  
+    if (!confirmed) {
+      return; // Si no confirma, no continúa con la eliminación
+    }
     this.auxService.ventanaCargando();
     this.usersService.Delete('delete-user', event.idUsuario).subscribe({
       next: async (data: any) => {
